update routes authorization on users controller

+++ backendAPI/app/Http/Controllers/UserController.php
@@ -35,7 +35,7 @@ class UserController extends Controller
                     $users = User::with('userInfo')->get();
 
 
-                        $users->each(function ($user) {
+                    $users->each(function ($user) {
                         $user->userInfo->profile_picture_path = Storage::disk('public')->url($user->userInfo->profile_picture_path);
                     });
 
@@ -64,53 +64,65 @@ class UserController extends Controller
      */
     public function store(UserStoreRequest $request)
     {
-        if ($request->get('role') != null) {
-            $role = Roles::where('id', $request->get('role'))->first();
-        } else {
-            $role = Roles::where('name', 'user')->first();
-        }
+        if (Auth::guard('api')->check()) { // Check if user is logged in
+            if (Auth::guard('api')->user()->hasRole('admin')) { // Check if user is admin
+                if ($request->get('role') != null) {
+                    $role = Roles::where('id', $request->get('role'))->first();
+                } else {
+                    $role = Roles::where('name', 'user')->first();
+                }
 
-        try {
+                try {
 
-            $validatedData = $request->validated();
+                    $validatedData = $request->validated();
 
-            try{
-                $user = new User([
-                    'name' => $validatedData['name'],
-                    'email' => $validatedData['email'],
-                    'password' => Hash::make($validatedData['password']),
-                    'internalcode' => $validatedData['internalcode'],
-                ]);
-            } catch (Exception $e) {
-                return response()->json($e->getMessage(), 500);
+                    try {
+                        $user = new User([
+                            'name' => $validatedData['name'],
+                            'email' => $validatedData['email'],
+                            'password' => Hash::make($validatedData['password']),
+                            'internalcode' => $validatedData['internalcode'],
+                        ]);
+                    } catch (Exception $e) {
+                        return response()->json($e->getMessage(), 500);
+                    }
+
+                    $user->setAttribute('normalized_name', strtoupper($validatedData['name']));
+
+                    $user->save();
+                    $user->roles()->attach($role);
+
+                    return response()->json($user->id, 200);
+
+                } catch (Exception $e) {
+
+                    return response()->json($e->getMessage(), 500);
+                }
+            } else {
+                return response()->json("Not authorized", 401);
             }
-
-            $user->setAttribute('normalized_name', strtoupper($validatedData['name']));
-
-            $user->save();
-            $user->roles()->attach($role);
-
-            return response()->json($user->id, 200);
-
-        } catch (Exception $e) {
-
-            return response()->json($e->getMessage(), 500);
+        } else {
+            return response()->json("Not logged in", 401);
         }
-
     }
 
     public function show(User $user)
     {
         if (Auth::guard('api')->check()) { // Check if user is logged in
-            try {
-                $user->load('userInfo', 'roles');
-                //converter data na formato d-m-Y
-                $user->userInfo->birthday_date = date('d-m-Y', strtotime($user->userInfo->birthday_date));
-                $user->userInfo->profile_picture_path = Storage::disk('public')->url($user->userInfo->profile_picture_path);
-                return response()->json($user, 200);
-            } catch (Exception $e) {
-                return response()->json($e, 500);
+            if (Auth::guard('api')->user()->hasRole('admin') || Auth::guard('api')->user()->id == $user->id) {
+                try {
+                    $user->load('userInfo', 'roles');
+                    //converter data na formato d-m-Y
+                    $user->userInfo->birthday_date = date('d-m-Y', strtotime($user->userInfo->birthday_date));
+                    $user->userInfo->profile_picture_path = Storage::disk('public')->url($user->userInfo->profile_picture_path);
+                    return response()->json($user, 200);
+                } catch (Exception $e) {
+                    return response()->json($e, 500);
+                }
+            } else {
+                return response()->json("Not authorized", 401);
             }
+
         } else {
             return response()->json("Not logged in", 401);
         }
@@ -126,23 +138,27 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         if (Auth::guard('api')->check()) { // Check if user is logged in
-            try {
-                $validatedData = $request->validate([
-                    'name' => 'required|max:255',
-                    'email' => 'required|max:255',
-                    'internalcode' => 'required|max:255',
-                ]);
+            if (Auth::guard('api')->user()->hasRole('admin') || Auth::guard('api')->user()->id == $user->id) {
+                try {
+                    $validatedData = $request->validate([
+                        'name' => 'required|max:255',
+                        'email' => 'required|max:255',
+                        'internalcode' => 'required|max:255',
+                    ]);
 
-                $roles = Roles::query()->where('id', $request->get('role'))->get();
+                    $roles = Roles::query()->where('id', $request->get('role'))->get();
 
 
-                $user->roles()->sync($roles);
-                \Log::info($user);
+                    $user->roles()->sync($roles);
+                    \Log::info($user);
 
-                $user->update($validatedData);
-                return response()->json($user, 200);
-            } catch (Exception $e) {
-                return response()->json($e, 500);
+                    $user->update($validatedData);
+                    return response()->json($user, 200);
+                } catch (Exception $e) {
+                    return response()->json($e, 500);
+                }
+            } else {
+                return response()->json("Not authorized", 401);
             }
 
         } else {
@@ -159,7 +175,7 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         if (Auth::guard('api')->check()) { // Check if user is logged in
-           // if (Auth::user()->hasRole('admin')) { // Check if user is admin
+            if (Auth::guard('api')->user()->hasRole('admin')) { // Check if user is admin
 
                 try {
                     $user->roles()->detach();
@@ -180,9 +196,9 @@ class UserController extends Controller
                     return response()->json($e->getMessage(), 500);
                 }
 
-           // } else {
-           //     return response()->json("Not authorized", 401);
-           // }
+            } else {
+                return response()->json("Not authorized", 401);
+            }
         } else {
             return response()->json("Not logged in", 401);
         }
@@ -219,8 +235,7 @@ class UserController extends Controller
         try {
             $user = Auth::guard('api')->user();
             return response()->json($user, 200);
-        }
-        catch (Exception $e) {
+        } catch (Exception $e) {
             return response()->json($e, 500);
         }
     }
@@ -241,25 +256,25 @@ class UserController extends Controller
 
     public function changePassword(UserChangePasswordRequest $request)
     {
-            try {
-                $validatedData = $request->validated();
-                $user = Auth::guard('api')->user();
+        try {
+            $validatedData = $request->validated();
+            $user = Auth::guard('api')->user();
 
-                if($validatedData['currentPassword'] == $validatedData['newPassword']){
-                    return response()->json("New password cannot be the same as the current password", 401);
-                }
-
-                if(!Hash::check($validatedData['currentPassword'], $user->password)){
-                    return response()->json("Current password is incorrect", 401);
-                }
-
-                $user->setAttribute('password', Hash::make($validatedData['newPassword']));
-
-                $user->save();
-                return response()->json($user, 200);
-            } catch (Exception $e) {
-                return response()->json($e, 500);
+            if ($validatedData['currentPassword'] == $validatedData['newPassword']) {
+                return response()->json("New password cannot be the same as the current password", 401);
             }
+
+            if (!Hash::check($validatedData['currentPassword'], $user->password)) {
+                return response()->json("Current password is incorrect", 401);
+            }
+
+            $user->setAttribute('password', Hash::make($validatedData['newPassword']));
+
+            $user->save();
+            return response()->json($user, 200);
+        } catch (Exception $e) {
+            return response()->json($e, 500);
+        }
     }
 
 }
